backend/requirements.txt
django>=5.2 
djangorestframework>=3.15 
psycopg2-binary>=2.9 
python-decouple>=3.8 
django-cors-headers>=4.4 
gunicorn>=23.0 
djangorestframework-simplejwt>=5.3
openai>=1.0
dj-database-url>=2.0
whitenoise==6.7.0
pytest
pytest-django
pytest-html
pytz